pyproject.toml
[tool.poetry]
name = "utility-bill-scraper"
version = "0.1.0"
description = "Utility bill scraper for extracting data from pdfs and websites."
authors = ["Ryan Fobel <ryan@fobel.net>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
numpy = "^1.21.2"
pandas = "^1.3.3"
selenium = "^3.141.0"
arrow = "^1.2.0"
pdfminer = "^20191125"
beautifulsoup4 = "^4.10.0"
matplotlib = "^3.4.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
python-dotenv = "^0.19.1"
jupyterlab = "^3.1.18"
jupytext = "^1.13.0"
flake8 = "^4.0.1"
black = "^21.9b0"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
